# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Client","description":"the Client Api"},{"name":"Product","description":"the Product Api"},{"name":"Ordine","description":"the Ordine Api"},{"name":"Bill","description":"the Bill Api"}],"paths":{"/api/products":{"get":{"tags":["Product"],"summary":"Fetch all products","description":"Fetches all products entities and their data from data source","operationId":"getAllProducts","responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}}}}},"put":{"tags":["Product"],"summary":"Update a product","description":"Update a product by a specific id from the list of products in the data source","operationId":"updateProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}},"required":true},"responses":{"404":{"description":"there is not product with the given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"200":{"description":"successfully updated a product","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}}},"post":{"tags":["Product"],"summary":"Adds a product","description":"Adds a product to the list of products in the data source","operationId":"saveProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDto"}}},"required":true},"responses":{"500":{"description":"invalid product cannot be added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}},"201":{"description":"successfully added a product","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}}}},"/api/ordini":{"get":{"tags":["Ordine"],"summary":"Fetch all ordini","description":"Fetches all ordini entities and their data from data source","operationId":"getAllOrdini","responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrdineDto"}}}}}}},"put":{"tags":["Ordine"],"summary":"Update a ordine","description":"Update a ordine by a specific id from the list of ordini in the data source","operationId":"updateOrdine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}},"required":true},"responses":{"404":{"description":"there is not ordine with the given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}}},"200":{"description":"successfully updated a ordine","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}}}}},"post":{"tags":["Ordine"],"summary":"Adds a ordine","description":"Adds a ordine to the list of ordini in the data source","operationId":"saveOrdine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}},"required":true},"responses":{"201":{"description":"successfully added a ordine","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}}},"500":{"description":"invalid ordine cannot be added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}}}}}},"/api/clients":{"get":{"tags":["Client"],"summary":"Fetch all clients","description":"Fetches all clients entities and their data from data source","operationId":"getAllClients","responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClientDto"}}}}}}},"put":{"tags":["Client"],"summary":"Update a client","description":"Update a client by a specific id from the list of clients in the data source","operationId":"updateClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}},"required":true},"responses":{"404":{"description":"there is not client with the given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"200":{"description":"successfully updated a client","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}}}},"post":{"tags":["Client"],"summary":"Adds a client","description":"Adds a client to the list of clients in the data source","operationId":"saveClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}},"required":true},"responses":{"201":{"description":"successfully added a client","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"500":{"description":"invalid client cannot be added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}}}}},"/api/bills":{"get":{"tags":["Bill"],"summary":"Fetch all bills","description":"Fetches all bills entities and their data from data source","operationId":"getAllBills","responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BillDto"}}}}}}},"put":{"tags":["Bill"],"summary":"Update a bill","description":"Update a bill by a specific id from the list of bills in the data source","operationId":"updateBill","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}},"required":true},"responses":{"200":{"description":"successfully updated a bill","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}}},"404":{"description":"there is not bill with the given id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}}}}},"post":{"tags":["Bill"],"summary":"Adds a bill","description":"Adds a bill to the list of bills in the data source","operationId":"saveBill","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}},"required":true},"responses":{"500":{"description":"invalid bill cannot be added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}}},"201":{"description":"successfully added a bill","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}}}}}},"/api/products/{id}":{"get":{"tags":["Product"],"summary":"Fetch a product by Id","description":"Fetches a product entity and their data from data source by an specific Id","operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDto"}}}}}},"delete":{"tags":["Product"],"summary":"Delete a product","description":"Delete a product by a specific id from the list of products in the data source","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"there is not product with the given id"},"202":{"description":"successfully deleted a product"}}}},"/api/ordini/{id}":{"get":{"tags":["Ordine"],"summary":"Fetch a ordine by Id","description":"Fetches a ordine entity and their data from data source by an specific Id","operationId":"getOrdineById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrdineDto"}}}}}},"delete":{"tags":["Ordine"],"summary":"Delete a ordine","description":"Delete a ordine by a specific id from the list of ordini in the data source","operationId":"deleteOrdine","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"there is not ordine with the given id"},"202":{"description":"successfully deleted a ordine"}}}},"/api/clients/{id}":{"get":{"tags":["Client"],"summary":"Fetch a client by Id","description":"Fetches a client entity and their data from data source by an specific Id","operationId":"getClientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}}}},"delete":{"tags":["Client"],"summary":"Delete a client","description":"Delete a client by a specific id from the list of clients in the data source","operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"there is not client with the given id"},"202":{"description":"successfully deleted a client"}}}},"/api/bills/{id}":{"get":{"tags":["Bill"],"summary":"Fetch a bill by Id","description":"Fetches a bill entity and their data from data source by an specific Id","operationId":"getBillById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BillDto"}}}}}},"delete":{"tags":["Bill"],"summary":"Delete a bill","description":"Delete a bill by a specific id from the list of bills in the data source","operationId":"deleteBill","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"successfully deleted a bill"},"404":{"description":"there is not bill with the given id"}}}}},"components":{"schemas":{"ClientDto":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":100,"minLength":1,"pattern":"^[a-zA-Z\\s]+$","type":"string"},"email":{"type":"string"},"ords":{"maxItems":2147483647,"minItems":0,"type":"array","items":{"$ref":"#/components/schemas/OrdineDto"}}}},"OrdineDto":{"required":["prods","total"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateOrdine":{"type":"string","format":"date-time"},"total":{"minimum":0,"type":"number","format":"double"},"cli":{"$ref":"#/components/schemas/ClientDto"},"prods":{"maxItems":2147483647,"minItems":0,"type":"array","items":{"$ref":"#/components/schemas/ProductDto"}}}},"ProductDto":{"required":["name","ords","price"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":100,"minLength":1,"type":"string"},"price":{"minimum":0,"type":"number","format":"double"},"ords":{"type":"array","items":{"$ref":"#/components/schemas/OrdineDto"}}}},"BillDto":{"required":["totalAmount"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dateBill":{"type":"string","format":"date-time"},"totalAmount":{"minimum":0,"type":"number","format":"double"},"ord":{"$ref":"#/components/schemas/OrdineDto"}}}}}}